rename editViewMode
