+++ public/defaultConfig.yaml
@@ -76,7 +76,5 @@ config:
     FEEDBACK:
       isEnabled: true
       link: https://sapinsights.eu.qualtrics.com/jfe/form/SV_d3UPNymSgUHAb9Y?product=SAP%20BTP,%20Kyma%20Runtime&product_filter=Kyma
-    EDIT_VIEW:
-      viewType: auto
     COLUMN_LAYOUT:
       isEnabled: true
